+++ Pasteleria/Pasteleria/src/main/java/com/example/demo/service/EmpleadoService.java
@@ -4,12 +4,8 @@ import com.example.demo.model.Empleado;
 import java.util.List;
 
 public interface EmpleadoService {
-
     List<Empleado> findAll();
-
-    Empleado findById(Long id);
-
+    Empleado findById(Integer id); 
     Empleado save(Empleado empleado);
-
-    void deleteById(Long id);
+    void deleteById(Integer id); 
 }
